Moved filters to a separate FilterComponent
