+++ src/App.js
@@ -4,7 +4,8 @@ import { Switch, Route } from "react-router-dom";
 
 import Header from "./components/Header";
 import SearchFilterBar from "./components/SearchFilterBar";
-import List from "./components/List";
+// import { List } from "./components/List";
+import { FilterComponent } from "./components/List/FilterComponent";
 import Restaurant from "./components/Restaurant";
 
 import "./index.scss";
@@ -32,7 +33,7 @@ function App() {
           </Route>
           <Route path="/">
             <SearchFilterBar />
-            <List rests={content} />
+            <FilterComponent rests={content} />
           </Route>
         </Switch>
       </main>
